+++ apps/block_scout_web/assets/webpack.config.js
@@ -51,6 +51,7 @@ const appJs =
       'address-logs': './js/pages/address/logs.js',
       'address-validations': './js/pages/address/validations.js',
       'validated-transactions': './js/pages/transactions.js',
+      'verified-contracts': './js/pages/verified_contracts.js',
       'pending-transactions': './js/pages/pending_transactions.js',
       'transaction': './js/pages/transaction.js',
       'verification-form': './js/pages/verification_form.js',