+++ apps/block_scout_web/lib/block_scout_web/chain.ex
@@ -25,6 +25,7 @@ defmodule BlockScoutWeb.Chain do
     InternalTransaction,
     Log,
     StakingPool,
+    SmartContract,
     Token,
     TokenTransfer,
     Transaction,
@@ -221,6 +222,10 @@ defmodule BlockScoutWeb.Chain do
     [paging_options: %{@default_paging_options | key: {name, type, value}}]
   end
 
+  def paging_options(%{"smart_contract_id" => id}) do
+    [paging_options: %{@default_paging_options | key: {id}}]
+  end
+
   def paging_options(_params), do: [paging_options: @default_paging_options]
 
   def put_key_value_to_paging_options([paging_options: paging_options], key, value) do
@@ -235,6 +240,11 @@ defmodule BlockScoutWeb.Chain do
       _ ->
         1
     end
+  end
+
+  def fetch_page_number(%{"items_count" => item_count_str}) do
+    {items_count, _} = Integer.parse(item_count_str)
+    div(items_count, @page_size) + 1
   end
 
   def fetch_page_number(_), do: 1
@@ -360,6 +370,10 @@ defmodule BlockScoutWeb.Chain do
 
   defp paging_params(%{address_hash: address_hash, total_gas: total_gas}) do
     %{"address_hash" => address_hash, "total_gas" => Decimal.to_integer(total_gas)}
+  end
+
+  defp paging_params(%SmartContract{} = smart_contract) do
+    %{"smart_contract_id" => smart_contract.id}
   end
 
   defp paging_params(%{